Refactor MyStoreStatsRequest
