+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/stats/datasource/StoreStatsRequest.kt
@@ -2,23 +2,14 @@ package com.woocommerce.android.ui.stats.datasource
 
 import com.woocommerce.commons.extensions.convertedFrom
 
-sealed class MyStoreStatsRequest {
+sealed class StoreStatsRequest {
     data class Data(
         private val revenueData: RevenueData?,
         private val visitorData: Int?
-    ) : MyStoreStatsRequest() {
-        val isFinished
-            get() = revenueData != null &&
-                visitorData != null
-        val revenue
-            get() = revenueData?.totalRevenue.orEmpty()
-
-        val ordersCount
-            get() = revenueData?.orderCount ?: 0
-
-        val visitorsCount
-            get() = visitorData ?: 0
-
+    ) : StoreStatsRequest() {
+        val revenue get() = revenueData?.totalRevenue.orEmpty()
+        val ordersCount get() = revenueData?.orderCount ?: 0
+        val visitorsCount get() = visitorData ?: 0
         val conversionRate: String
             get() {
                 val ordersCount = revenueData?.orderCount ?: 0
@@ -26,11 +17,15 @@ sealed class MyStoreStatsRequest {
                 return ordersCount convertedFrom visitorsCount
             }
 
+        val isFinished
+            get() = revenueData != null &&
+                visitorData != null
+
         data class RevenueData(
             val totalRevenue: String,
             val orderCount: Int
         )
     }
 
-    data object Error : MyStoreStatsRequest()
+    data object Error : StoreStatsRequest()
 }
